refactor: infra 안의 Client 를 domain 에 맞추어 이동
KitchenridersClient 는 order 하위의 infra 패키지로,
PurgomalumClient 는 common 하위의 infra 패키지로 이동
