+++ src/main/java/kitchenpos/common/infra/DefaultPurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.common.infra;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
@@ -18,9 +18,9 @@ public class DefaultPurgomalumClient implements PurgomalumClient {
     @Override
     public boolean containsProfanity(final String text) {
         final URI url = UriComponentsBuilder.fromUriString("https://www.purgomalum.com/service/containsprofanity")
-            .queryParam("text", text)
-            .build()
-            .toUri();
+                .queryParam("text", text)
+                .build()
+                .toUri();
         return Boolean.parseBoolean(restTemplate.getForObject(url, String.class));
     }
 }
