+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.order.application;
 
-import kitchenpos.infra.FakeKitchenridersClient;
+import kitchenpos.order.infra.FakeKitchenridersClient;
 import kitchenpos.menu.domain.InMemoryMenuRepository;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.*;
